Storage Views

+++ database/migrations/2023_01_06_135200_add_source_sort_to_articles.php
@@ -14,7 +14,7 @@ class AddSourceSortToArticles extends Migration
     public function up()
     {
         Schema::table('articles', function (Blueprint $table) {
-            $table->unsignedSmallInteger('source_sort')->nullable()->after('source_id');
+            $table->unsignedInteger('source_sort')->nullable()->after('source_id');
         });
     }
